AbstractNeuron don't depend on AbstractLayer anymore. AbstractNeuron::calculateGradient (and overrides) were rewritten to receive neuron data through lambda-function instead.
As of new design, the function receives a lambda expression returning the sum of the calculations on each neuron involved, where a calculation is passed as another lambda-parameter to the sum_function lambda.
User-lambda calls unit-wise calculation lambda for each neuron and sums the result. They are returned from it back to neuron in order to calculate final gradient.

The system is way more complicated but the it gives the advantage of independence from leyer-based structure. Now it'd be easier to implement randomly-structured neural network based on this Neurons.